Support running base testsuite on Windows
Closes #12648

+++ testsuite/integration-arquillian/tests/base/src/main/java/org/keycloak/testsuite/arquillian/containers/KeycloakQuarkusServerDeployableContainer.java
@@ -14,21 +14,26 @@ import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URISyntaxException;
 import java.net.URL;
+import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
+import java.nio.file.LinkOption;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardCopyOption;
+import java.nio.file.attribute.BasicFileAttributes;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
 import java.security.cert.X509Certificate;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.stream.Collectors;
 
 import org.apache.commons.exec.StreamPumper;
-import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.SystemUtils;
 import org.jboss.arquillian.container.spi.client.container.DeployableContainer;
 import org.jboss.arquillian.container.spi.client.container.DeploymentException;
@@ -48,6 +53,7 @@ import org.keycloak.testsuite.arquillian.SuiteContext;
  */
 public class KeycloakQuarkusServerDeployableContainer implements DeployableContainer<KeycloakQuarkusConfiguration> {
 
+    private static final int DEFAULT_SHUTDOWN_TIMEOUT_SECONDS = 10;
     private static final String AUTH_SERVER_QUARKUS_MAP_STORAGE_PROFILE = "auth.server.quarkus.mapStorage.profile";
 
     private static final Logger log = Logger.getLogger(KeycloakQuarkusServerDeployableContainer.class);
@@ -87,11 +93,15 @@ public class KeycloakQuarkusServerDeployableContainer implements DeployableConta
 
     @Override
     public void stop() throws LifecycleException {
-        container.destroy();
-        try {
-            container.waitFor(10, TimeUnit.SECONDS);
-        } catch (InterruptedException e) {
-            container.destroyForcibly();
+        if (container.isAlive()) {
+            try {
+                destroyDescendantsOnWindows(container, false);
+                container.destroy();
+                container.waitFor(10, TimeUnit.SECONDS);
+            } catch (InterruptedException e) {
+                destroyDescendantsOnWindows(container, true);
+                container.destroyForcibly();
+            }
         }
     }
 
@@ -118,6 +128,10 @@ public class KeycloakQuarkusServerDeployableContainer implements DeployableConta
     public void undeploy(Archive<?> archive) throws DeploymentException {
         File wrkDir = configuration.getProvidersPath().resolve("providers").toFile();
         try {
+            if (isWindows()) {
+                // stop before updating providers to avoid file locking issues on Windows
+                stop();
+            }
             Files.deleteIfExists(wrkDir.toPath().resolve(archive.getName()));
             restartServer();
         } catch (Exception e) {
@@ -172,7 +186,7 @@ public class KeycloakQuarkusServerDeployableContainer implements DeployableConta
         builder.environment().put("KEYCLOAK_ADMIN_PASSWORD", "admin");
 
         if (restart.compareAndSet(false, true)) {
-            FileUtils.deleteDirectory(configuration.getProvidersPath().resolve("data").toFile());
+            deleteDirectory(configuration.getProvidersPath().resolve("data"));
         }
 
         return builder.start();
@@ -346,7 +360,11 @@ public class KeycloakQuarkusServerDeployableContainer implements DeployableConta
         additionalBuildArgs = Collections.emptyList();
     }
 
-    private void deployArchiveToServer(Archive<?> archive) throws IOException {
+    private void deployArchiveToServer(Archive<?> archive) throws IOException, LifecycleException {
+        if (isWindows()) {
+            // stop before updating providers to avoid file locking issues on Windows
+            stop();
+        }
         File providersDir = configuration.getProvidersPath().resolve("providers").toFile();
         InputStream zipStream = archive.as(ZipExporter.class).exportAsInputStream();
         Files.copy(zipStream, providersDir.toPath().resolve(archive.getName()), StandardCopyOption.REPLACE_EXISTING);
@@ -357,9 +375,9 @@ public class KeycloakQuarkusServerDeployableContainer implements DeployableConta
         start();
     }
 
-    private static String getCommand() {
-        if (SystemUtils.IS_OS_WINDOWS) {
-            return "kc.bat";
+    private String getCommand() {
+        if (isWindows()) {
+            return configuration.getProvidersPath().resolve("bin").resolve("kc.bat").toString();
         }
         return "./kc.sh";
     }
@@ -370,5 +388,66 @@ public class KeycloakQuarkusServerDeployableContainer implements DeployableConta
 
     public void setAdditionalBuildArgs(List<String> newArgs) {
         additionalBuildArgs = newArgs;
+    }
+
+    private void destroyDescendantsOnWindows(Process parent, boolean force) {
+        if (!isWindows()) {
+            return;
+        }
+
+        CompletableFuture allProcesses = CompletableFuture.completedFuture(null);
+
+        for (ProcessHandle process : parent.descendants().collect(Collectors.toList())) {
+            if (force) {
+                process.destroyForcibly();
+            } else {
+                process.destroy();
+            }
+
+            allProcesses = CompletableFuture.allOf(allProcesses, process.onExit());
+        }
+
+        try {
+            allProcesses.get(DEFAULT_SHUTDOWN_TIMEOUT_SECONDS, TimeUnit.SECONDS);
+        } catch (Exception cause) {
+            throw new RuntimeException("Failed to terminate descendants processes", cause);
+        }
+
+        try {
+            // TODO: remove this. do not ask why, but on Windows we are here even though the process was previously terminated
+            // without this pause, tests re-installing dist before tests should fail
+            // looks like pausing the current thread let windows to cleanup processes?
+            // more likely it is env dependent
+            Thread.sleep(500);
+        } catch (InterruptedException e) {
+        }
+    }
+
+    private static boolean isWindows() {
+        return SystemUtils.IS_OS_WINDOWS;
+    }
+
+    public static void deleteDirectory(final Path directory) throws IOException {
+        if (Files.isDirectory(directory, new LinkOption[0])) {
+            Files.walkFileTree(directory, new SimpleFileVisitor<Path>() {
+                public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+                    try {
+                        Files.delete(file);
+                    } catch (IOException var4) {
+                    }
+
+                    return FileVisitResult.CONTINUE;
+                }
+
+                public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
+                    try {
+                        Files.delete(dir);
+                    } catch (IOException var4) {
+                    }
+
+                    return FileVisitResult.CONTINUE;
+                }
+            });
+        }
     }
 }
